% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaGridMean.R
\name{alaGridMean}
\alias{alaGridMean}
\title{ALA Layer Means per Grid}
\usage{
alaGridMean(data, km, ncol1, ncol2)
}
\arguments{
\item{data}{ALA dataframe containing point occurrence data and layer values}

\item{km}{Numeric. Distance of grid size in km^2 to be sampled}

\item{ncol1}{The column number corresponding to the first layer. Check with colnames(data)}

\item{ncol2}{The column number corresponding to the last layer. Check with colnames(data)}
}
\description{
Calculates individual cell means of layer variables, downloaded from ALA.
Values are grouped by species.
}
\examples{
df <- read.csv("example_input.csv")
x <- alaGridMean(df, 10, 38, 42)

}
