% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaSpeciesMean.R
\name{alaSpeciesMean}
\alias{alaSpeciesMean}
\title{ALA Layer Means by Species}
\usage{
alaSpeciesMean(data, km, ncol1, ncol2)
}
\arguments{
\item{data}{ALA dataframe containing point occurrence data and layer values}

\item{km}{Numeric. Distance of grid size in km^2 to be sampled}

\item{ncol1}{The column number corresponding to the first layer. Check with colnames(data)}

\item{ncol2}{The column number corresponding to the last layer. Check with colnames(data)}
}
\description{
Calculates the species means of layer variables, downloaded from ALA. Means
are calculated as per the specified grid size in km^2.
}
\examples{
df <- read.csv("example_input.csv")
x <- alaSpeciesMean(df, 10, 38, 42)

}
