% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eucs}
\alias{eucs}
\title{Eucalyptus point occurrence data with environmental layers}
\format{A data frame with 39904 rows and 8 variables}
\usage{
data(eucs)
}
\description{
Point data of 40 Eucalyptus spp. with 5 environmental layers prepared
and downloaded using Atlas of Australia's spatial portal.
}
\keyword{datasets}
