% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMean.R
\name{gridMean}
\alias{gridMean}
\title{Grid-based means of species point occurrences}
\usage{
gridMean(data, km)
}
\arguments{
\item{data}{Dataframe of point occurrence data and corresponding variables. 
First three columns must include species names, latitude and longitudes. All 
additional columns will contain the variables}

\item{km}{Numeric. Distance of grid size in km^2 to be applied}
}
\description{
Calculates species grid means of point occurrence-associated variables
(i.e. climate layers) of Australian distributions. Means are calculated 
based on the user-specified grid size (km^2).
}
\examples{
data(eucs)
grMean <- gridMean(eucs, 10)

}
