% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs}
\alias{gs}
\title{Mean genome size of Eucalypts}
\format{A data frame with 39904 rows and 8 variables}
\usage{
data(gs)
}
\description{
This dataset was derived from \url{https://github.com/eleanorbeavan/
euc_rates/blob/master/data/genome_size.csv} containing raw genome size
measurements of 40 Eucalyptus spp. (units pg/2C).
}
\keyword{datasets}
