% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpeciesTrait.R
\name{plotSpeciesTrait}
\alias{plotSpeciesTrait}
\title{Distribution of Species-Level Traits}
\usage{
plotSpeciesTrait(points, traits, trait_name, km)
}
\arguments{
\item{points}{Dataframe containing species point occurrence data. First
three columns must be ordered by species names, latitudes and longitudes.}

\item{traits}{Two column dataframe containing species names and trait means.}

\item{trait_name}{Character string specifying the trait name and
corresponding units (if applicable). This will be incorporated in the graph title.}

\item{km}{Numeric. Distance of grid size in km^2 to be sampled.}
}
\description{
Plots the distribution of Australian species-level traits. Three graphs are
generated (trait means, standard deviation and species richness), written as
one .png file. Values are calculated as per the specified grid size in km^2.
}
\examples{
# Plot the distribution of a subset of Eucalyptus species' genome sizes
pts <- read.csv("example_input.csv")
pts <- pts[,c(4,20,21)]
gs <- read.csv("gs_input.csv")
plotSpeciesTrait(pts, gs, "genome size (pg/2C)", 100)

}
