% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesMean.R
\name{speciesMean}
\alias{speciesMean}
\title{Grid-based species means of point occurrences}
\usage{
speciesMean(data, km)
}
\arguments{
\item{data}{Dataframe of point occurrence data and corresponding variables. 
First three columns must include species names, latitude and longitudes. All 
additional columns will contain the variables}

\item{km}{Numeric. Distance of grid size in km^2 to be applied}
}
\description{
Calculates species-level means of point occurrence-associated variables
(i.e. climate layers). Means are calculated based on the user-specified grid
size (km^2).
}
\examples{
df <- read.csv("example_input.csv")
x <- speciesMean(df, 10)

}
