% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaGridMean.R
\name{alaGridMean}
\alias{alaGridMean}
\title{ALA Layer Means by Species}
\usage{
alaGridMean(data, km)
}
\arguments{
\item{data}{ALA dataframe containing point occurrence data and layer values}

\item{km}{Numeric. Distance of grid size in km^2 to be sampled}
}
\description{
Calculates the mean of a layer variable per specified grid size (km^2) of
compiled species downloaded from ALA.
}
\examples{
alaGridMean(df, 10)
alaGridMean(data = df, km = 10)

}
